pyproject.toml
[project]
name = "prosper-challenge"
version = "0.1.0"
description = "Template repository for clinic appointment scheduling voice agent challenge"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[webrtc,daily,silero,elevenlabs,openai,local-smart-turn-v3,runner]",
    "pipecat-ai-cli",
    "httpx",
    "pydantic-settings",
    "playwright",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "pyright>=1.1.404,<2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.12.11,<1",
]

[tool.pytest.ini_options]
markers = [
    "unit: fast tests with no external dependencies",
    "integration: requires real Healthie credentials (deselected by default)",
]
asyncio_mode = "strict"

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]
